f.1 - Изменение типа данных денежных величин

+++ Vault/VaultDataModels/Models/IAccount.cs
@@ -6,6 +6,6 @@
 
         string Purpose { get; }
 
-        double Balance { get; }
+		decimal Balance { get; }
     }
 }
